support esm, cjs both
