+++ integrations/react/package.json
@@ -9,21 +9,24 @@
   "license": "MIT",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
+      "types": "./dist/esm/index.d.ts",
+      "require": "./dist/cjs/index.js",
+      "import": "./dist/esm/index.mjs"
     },
     "./stable": {
-      "types": "./dist/stable/index.d.ts",
-      "import": "./dist/stable/index.js"
+      "types": "./dist/esm/stable/index.d.ts",
+      "require": "./dist/cjs/stable/index.js",
+      "import": "./dist/esm/stable/index.mjs"
     },
     "./future": {
-      "types": "./dist/future/index.d.ts",
-      "import": "./dist/future/index.js"
+      "types": "./dist/esm/future/index.d.ts",
+      "require": "./dist/cjs/future/index.js",
+      "import": "./dist/esm/future/index.mjs"
     }
   },
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm/index.mjs",
+  "types": "./dist/esm/index.d.ts",
   "files": [
     "dist",
     "src",
@@ -31,7 +34,7 @@
   ],
   "scripts": {
     "build": "yarn build:js && yarn build:dts",
-    "build:dts": "tsc --emitDeclarationOnly",
+    "build:dts": "tsc --emitDeclarationOnly --outDir dist/esm && tsc --emitDeclarationOnly --outDir dist/cjs",
     "build:js": "node ./esbuild.config.js",
     "clean": "rimraf dist",
     "dev": "yarn build:js --watch && yarn build:dts --watch",
@@ -48,6 +51,7 @@
     "@stackflow/esbuild-config": "^1.0.3",
     "@types/react": "^18.3.3",
     "esbuild": "^0.23.0",
+    "esbuild-plugin-file-path-extensions": "^2.1.2",
     "react": "^18.3.1",
     "rimraf": "^3.0.2",
     "typescript": "^5.5.3"
