+++ integrations/react/src/future/StepActions.ts
@@ -1,0 +1,5 @@
+export type StepActions<ActivityParams> = {
+  pushStep: (params: ActivityParams, options?: {}) => void;
+  replaceStep: (params: ActivityParams, options?: {}) => void;
+  popStep: (options?: {}) => void;
+};
