+++ integrations/react/src/future/index.ts
@@ -1,12 +1,23 @@
-export * from "../__internal__/activity/useActivity";
-export * from "../__internal__/stack/useStack";
+/**
+ * Stack
+ */
+export * from "./stack";
+
+/**
+ * Types
+ */
 export * from "../__internal__/StackflowReactPlugin";
 export * from "./ActivityComponentType";
 export * from "./StackComponentType";
-export * from "./stack";
-export * from "./makeActions";
-export * from "./makeStepActions";
-export * from "./loader/useLoaderData";
+export * from "./Actions";
+export * from "./StepActions";
+
+/**
+ * Hooks
+ */
+export * from "../__internal__/stack/useStack";
+export * from "../__internal__/activity/useActivity";
 export * from "./useActivityParams";
+export * from "./loader/useLoaderData";
 export * from "./useFlow";
 export * from "./useStepFlow";
