+++ integrations/react/src/future/useFlow.ts
@@ -1,5 +1,6 @@
 import { useCoreActions } from "../__internal__/core";
-import { type Actions, makeActions } from "./makeActions";
+import type { Actions } from "./Actions";
+import { makeActions } from "./makeActions";
 
 export type FlowOutput = {
   useFlow: () => Actions;
