+++ integrations/react/src/future/useStepFlow.ts
@@ -3,7 +3,8 @@ import type {
   RegisteredActivityParamTypes,
 } from "@stackflow/config";
 import { useCoreActions } from "../__internal__/core";
-import { type StepActions, makeStepActions } from "./makeStepActions";
+import type { StepActions } from "./StepActions";
+import { makeStepActions } from "./makeStepActions";
 
 export function useStepFlow<
   ActivityName extends Extract<keyof RegisteredActivityParamTypes, string>,
